Implemented Adding 'Ordonnance' from Home screen with Patient selection.
Displaying the attached 'Ordonnance' list in PatientDetails screen.

Related work items: #92, #31

+++ liber-medical/Models/Ordonnance.cs
@@ -6,7 +6,7 @@ namespace libermedical.Models
     public class Ordonnance : BaseDTO
     {
 
-        public int Reference { set; get; }
+        public long Reference { set; get; }
         public DateTime AddDate { set; get; }
         public Patient Patient { set; get; }
         public StatusEnum Status { set; get; }
